third_party/WebKit/LayoutTests/inspector/sources/debugger-ui/async-call-stack-async-function.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>
async function foo()
{
    await Promise.resolve(1);
    await Promise.resolve(2);
    debugger;
}

async function boo()
{
    await Promise.resolve(3);
    await foo();
}

async function testFunction()
{
    await Promise.resolve(4);
    await boo();
}
//# sourceURL=test.js
</script>
<script>
var test = function()
{
    TestRunner.DebuggerAgent.setAsyncCallStackDepth(200);

    SourcesTestRunner.startDebuggerTestPromise(/* quiet */ true)
        .then(() => SourcesTestRunner.runTestFunctionAndWaitUntilPausedPromise())
        .then(() => dumpCallStackSidebarPane())
        .then(() => SourcesTestRunner.completeDebuggerTest());

    function dumpCallStackSidebarPane()
    {
        var pane = self.runtime.sharedInstance(Sources.CallStackSidebarPane);
        for (var element of pane.contentElement.querySelectorAll('.call-frame-item'))
            TestRunner.addResult(element.deepTextContent().replace(/VM\d+/g, "VM"));
    }
}

</script>
</head>
<body onload="runTest()">
<p>Tests that call stack sidebar contains correct labels for async await functions.</p>
</body>
</html>
